Remove search
At the moment, there is no search form implemented, and the
`/docs/search` URL route doesn't seem to be hit (404).

But also, I'm not sure /docs/search is a good idea and needs more
consideration. I think it's confusing for `/docs` to have search
when the site itself doesn't.

If we want search for docs, I would think we should just add a site-wide
search, which would then automatically search docs.

But I also think it would be reasonable to treat this as out of scope
for simply getting `/docs` live - we can do it quickly afterwards.